Subjects

+++ app/Http/Controllers/StudentController.php
@@ -4,10 +4,12 @@ namespace App\Http\Controllers;
 
 use App\Models\Student;
 use App\Http\Requests\StudentRequest;
+use App\Models\User;
 use App\Models\Program;
 use App\Models\Subject;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
+use Spatie\Permission\Models\Role;
 
 class StudentController extends Controller
 {
@@ -16,7 +18,7 @@ class StudentController extends Controller
      */
     public function index()
     {
-        $students = Student::with('program')->paginate(5);
+        $students = User::role('Student')->with(['student.program'])->paginate(5);
 
         if (request()->wantsJson()) {
             return response()->json($students);
@@ -31,19 +33,54 @@ class StudentController extends Controller
     public function create()
     {
         $programs = Program::all()->pluck('name', 'id');
-        return inertia('Students/Create', ['programs' => $programs]);
+        $subjects = Subject::all()->pluck('name', 'id');
+        return inertia('Students/Create', [
+            'programs' => $programs,
+            'subjects' => $subjects
+        ]);
     }
 
     /**
-     * 
+     *
      * Store a newly created resource in storage.
      * @param App\Http\Requests\StudentRequest
      * @return \illuminate\Http\Response
      */
     public function store(StudentRequest $request)
     {
-        $validatedData = $request->validated();
-        Student::create($validatedData);
+        $validatedData = $request->validate([
+            'name' => 'required|string',
+            'email' => 'required|email|unique:users,email',
+            'password' => 'required|string|min:6',
+            'document' => 'required|string',
+            'first_name' => 'required|string',
+            'last_name' => 'required|string',
+            'phone' => 'nullable|string',
+            'address' => 'nullable|string',
+            'city' => 'nullable|string',
+            'semester' => 'nullable|integer',
+            'program_id' => 'required|exists:programs,id',
+        ]);
+
+        $user = User::create([
+            'name' => $validatedData['name'],
+            'email' => $validatedData['email'],
+            'password' => bcrypt($validatedData['password']),
+        ]);
+
+        $user->assignRole('Student');
+
+        $user->student()->create([
+            'document' => $validatedData['document'],
+            'first_name' => $validatedData['first_name'],
+            'last_name' => $validatedData['last_name'],
+            'phone' => $validatedData['phone'],
+            'address' => $validatedData['address'],
+            'city' => $validatedData['city'],
+            'semester' => $validatedData['semester'],
+            'program_id' => $validatedData['program_id'],
+        ]);
+
         return redirect()->route('students.index');
     }
 
@@ -58,9 +95,10 @@ class StudentController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Student $student)
+    public function edit(User $student)
     {
         // Obtén la lista de programas
+        $student->load('student.program');
         $programs = Program::all()->pluck('name', 'id');
 
         return inertia('Students/Edit', [
@@ -72,10 +110,21 @@ class StudentController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(StudentRequest $request, Student $student)
+    public function update(StudentRequest $request, $user_id)
     {
+        // Obtén el usuario y carga la relación con student
+        $user = User::with('student')->findOrFail($user_id);
+
+        //Validar datos recibidos
         $validatedData = $request->validated();
-        $student->update($validatedData);
+        // Actualiza el estudiante relacionado
+        $user->student->update($validatedData);
+
+        // Actualizar también el email en la tabla users, si aplica
+        if (isset($validatedData['email'])) {
+            $user->update(['email' => $validatedData['email']]);
+        }
+
         return redirect()->route('students.index');
     }
 
@@ -83,9 +132,9 @@ class StudentController extends Controller
      * @param Program $program
      * Remove the specified resource from storage.
      */
-    public function destroy(Student $student)
+    public function destroy(User  $student)
     {
-        $student->delete();
+        $student->student()->delete();
         return redirect()->route('students.index');
     }
 
@@ -156,6 +205,14 @@ class StudentController extends Controller
             Log::info('Unassigning subject...');
 
             $student = Student::findOrFail($studentId);
+            $totalCredits = $student->subjects->sum('credits');
+            $subject = Subject::findOrFail($subjectId);
+            $subjectCredits = $subject->credits;
+
+            if (($totalCredits - $subjectCredits) < 7) {
+                return response()->json(['error' => 'Cannot unassign this subject. The student must have at least 7 credits assigned.'], 422);
+            }
+
             $student->subjects()->detach($subjectId);
 
             Log::info('Subject unassigned successfully.');
